jina-clip/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pillow==10.1.0
torch==2.1.0
transformers==4.36.0
requests==2.31.0
pydantic==2.0.0